# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Grey project API",
        "contact": {},
        "version": "dev"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/auth/sign-up": {
            "post": {
                "description": "Sign up",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sign up",
                "parameters": [
                    {
                        "description": "UserSignUpInput",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserSignUpInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/cart/": {
            "get": {
                "description": "Show list of items for user with hard coded userId (handler.go -\u003e const)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show cart for user",
                "responses": {}
            }
        },
        "/api/cart/add/": {
            "post": {
                "description": "Add product to cart for user with hard coded userId (handler.go -\u003e const)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add product to cart for user",
                "parameters": [
                    {
                        "description": "CartInput",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CartInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/cart/checkout": {
            "get": {
                "description": "Checkout order for user with hard coded userId (handler.go -\u003e const)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Checkout order for user",
                "responses": {}
            }
        },
        "/api/order/": {
            "get": {
                "description": "Show list of orders for user with hard coded userId (handler.go -\u003e const)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show orders for user",
                "responses": {}
            }
        },
        "/api/order/{order_id}": {
            "get": {
                "description": "Show order details for user with hard coded userId(handler.go -\u003e const)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show order detail for user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "int",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/product/": {
            "get": {
                "description": "Show list of products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show list of products",
                "responses": {}
            }
        },
        "/api/product/create": {
            "post": {
                "description": "Create product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create product",
                "parameters": [
                    {
                        "description": "ProductInput",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ProductInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/product/{product_id}": {
            "get": {
                "description": "Get product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "int",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/product/{product_id}/update-price": {
            "post": {
                "description": "Update product price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update product price",
                "parameters": [
                    {
                        "description": "ProductPriceInput",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ProductPriceInput"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "int",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "domain.CartInput": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "domain.ProductInput": {
            "type": "object",
            "required": [
                "name",
                "price",
                "quantity"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "minimum": 1
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "domain.ProductPriceInput": {
            "type": "object",
            "required": [
                "price"
            ],
            "properties": {
                "price": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "domain.UserSignUpInput": {
            "type": "object",
            "required": [
                "age",
                "firstname",
                "is_married",
                "lastname",
                "password",
                "username"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "minimum": 18
                },
                "firstname": {
                    "type": "string"
                },
                "is_married": {
                    "type": "boolean"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}